cleaning

+++ configs/prame/install.sh
@@ -54,9 +54,14 @@ EOF
 }
 
 install_google_chrome() {
-  wget https://dl-ssl.google.com/linux/linux_signing_key.pub -O /tmp/google.pub
-  sudo gpg --no-default-keyring --keyring /etc/apt/keyrings/google-chrome.gpg --import /tmp/google.pub
-  echo 'deb [arch=amd64 signed-by=/etc/apt/keyrings/google-chrome.gpg]  https://dl.google.com/linux/chrome/deb/ stable main' | sudo tee /etc/apt/sources.list.d/google-chrome.list
+  if apt policy google-chrome-stable 2>/dev/null
+  then
+    echo 'Repo exists'
+  else
+    wget https://dl-ssl.google.com/linux/linux_signing_key.pub -O /tmp/google.pub
+    sudo gpg --no-default-keyring --keyring /etc/apt/keyrings/google-chrome.gpg --import /tmp/google.pub
+    echo 'deb [arch=amd64 signed-by=/etc/apt/keyrings/google-chrome.gpg]  https://dl.google.com/linux/chrome/deb/ stable main' | sudo tee /etc/apt/sources.list.d/google-chrome.list
+  fi
 
   sudo apt update
   sudo apt install google-chrome-stable -y
@@ -64,44 +69,54 @@ install_google_chrome() {
 }
 
 install_1password() {
-  curl -sS https://downloads.1password.com/linux/keys/1password.asc | sudo gpg --batch --yes --dearmor --output /usr/share/keyrings/1password-archive-keyring.gpg
-
-  echo 'deb [arch=amd64 signed-by=/usr/share/keyrings/1password-archive-keyring.gpg] https://downloads.1password.com/linux/debian/amd64 stable main' | sudo tee /etc/apt/sources.list.d/1password.list
-
-  sudo mkdir -p /etc/debsig/policies/AC2D62742012EA22/
-
-  curl -sS https://downloads.1password.com/linux/debian/debsig/1password.pol | sudo tee /etc/debsig/policies/AC2D62742012EA22/1password.pol
-
-  sudo mkdir -p /usr/share/debsig/keyrings/AC2D62742012EA22
-
-  curl -sS https://downloads.1password.com/linux/keys/1password.asc | sudo gpg --batch --yes --dearmor --output /usr/share/debsig/keyrings/AC2D62742012EA22/debsig.gpg
+  if apt policy 1password 2>/dev/null
+  then
+    echo 'Repo exists'
+  else
+    curl -sS https://downloads.1password.com/linux/keys/1password.asc | sudo gpg --batch --yes --dearmor --output /usr/share/keyrings/1password-archive-keyring.gpg
+    echo 'deb [arch=amd64 signed-by=/usr/share/keyrings/1password-archive-keyring.gpg] https://downloads.1password.com/linux/debian/amd64 stable main' | sudo tee /etc/apt/sources.list.d/1password.list
+    sudo mkdir -p /etc/debsig/policies/AC2D62742012EA22/
+    curl -sS https://downloads.1password.com/linux/debian/debsig/1password.pol | sudo tee /etc/debsig/policies/AC2D62742012EA22/1password.pol
+    sudo mkdir -p /usr/share/debsig/keyrings/AC2D62742012EA22
+    curl -sS https://downloads.1password.com/linux/keys/1password.asc | sudo gpg --batch --yes --dearmor --output /usr/share/debsig/keyrings/AC2D62742012EA22/debsig.gpg
+  fi
 
   sudo apt update
   sudo apt install 1password -y
 }
 
 install_terraform() {
-  wget -O- https://apt.releases.hashicorp.com/gpg | sudo gpg --batch --yes --dearmor -o /usr/share/keyrings/hashicorp-archive-keyring.gpg
-
-  echo "deb [signed-by=/usr/share/keyrings/hashicorp-archive-keyring.gpg] https://apt.releases.hashicorp.com $(lsb_release -cs) main" | sudo tee /etc/apt/sources.list.d/hashicorp.list
+  if apt policy terraform 2>/dev/null
+  then
+    echo 'Repo exists'
+  else
+    wget -O- https://apt.releases.hashicorp.com/gpg | sudo gpg --batch --yes --dearmor -o /usr/share/keyrings/hashicorp-archive-keyring.gpg
+    echo "deb [signed-by=/usr/share/keyrings/hashicorp-archive-keyring.gpg] https://apt.releases.hashicorp.com $(lsb_release -cs) main" | sudo tee /etc/apt/sources.list.d/hashicorp.list
+  fi
 
   sudo apt update
   sudo apt install terraform -y
 }
 
 install_docker() {
-  sudo apt-get update
-  sudo apt-get install -y ca-certificates curl
-  sudo install -m 0755 -d /etc/apt/keyrings
-  sudo curl -fsSL https://download.docker.com/linux/ubuntu/gpg -o /etc/apt/keyrings/docker.asc
-  sudo chmod a+r /etc/apt/keyrings/docker.asc
+  if apt policy docker-ce 2>/dev/null
+  then
+    echo 'Repo exists'
+  else
+    sudo apt update
+    sudo apt install -y ca-certificates curl
+    sudo install -m 0755 -d /etc/apt/keyrings
+    sudo curl -fsSL https://download.docker.com/linux/ubuntu/gpg -o /etc/apt/keyrings/docker.asc
+    sudo chmod a+r /etc/apt/keyrings/docker.asc
 
-  echo \
-    "deb [arch=$(dpkg --print-architecture) signed-by=/etc/apt/keyrings/docker.asc] https://download.docker.com/linux/ubuntu \
-    $(. /etc/os-release && echo "$VERSION_CODENAME") stable" | \
-    sudo tee /etc/apt/sources.list.d/docker.list > /dev/null
-  sudo apt-get update
-  sudo apt-get install -y \
+    echo \
+      "deb [arch=$(dpkg --print-architecture) signed-by=/etc/apt/keyrings/docker.asc] https://download.docker.com/linux/ubuntu \
+      $(. /etc/os-release && echo "$VERSION_CODENAME") stable" | \
+      sudo tee /etc/apt/sources.list.d/docker.list > /dev/null
+  fi
+
+  sudo apt update
+  sudo apt install -y \
     docker-ce \
     docker-ce-cli \
     containerd.io \
@@ -113,28 +128,44 @@ install_docker() {
 }
 
 install_vscode() {
-  sudo apt-get install wget gpg
-  wget -qO- https://packages.microsoft.com/keys/microsoft.asc | gpg --batch --yes --dearmor > microsoft.gpg
-  sudo install -D -o root -g root -m 644 microsoft.gpg /usr/share/keyrings/microsoft.gpg
-  echo "deb [arch=amd64,arm64,armhf signed-by=/usr/share/keyrings/microsoft.gpg] https://packages.microsoft.com/repos/code stable main" | sudo tee /etc/apt/sources.list.d/vscode.list > /dev/null
-  rm -f microsoft.gpg
+  if apt policy code 2>/dev/null
+  then
+    echo 'Repo exists'
+  else
+    sudo apt install wget gpg apt-transport-https 
+    wget -qO- https://packages.microsoft.com/keys/microsoft.asc | gpg --batch --yes --dearmor > microsoft.gpg
+    sudo install -D -o root -g root -m 644 microsoft.gpg /usr/share/keyrings/microsoft.gpg
+    echo "deb [arch=amd64,arm64,armhf signed-by=/usr/share/keyrings/microsoft.gpg] https://packages.microsoft.com/repos/code stable main" | sudo tee /etc/apt/sources.list.d/vscode.list > /dev/null
+    rm -f microsoft.gpg
+  fi
 
   sudo apt update
-  sudo apt install -y apt-transport-https code 
+  sudo apt install -y code 
 }
 
 install_nordvpn() {
-  sh <(wget -qO - https://downloads.nordcdn.com/apps/linux/install.sh)
+  if apt policy nordvpn 2>/dev/null
+  then
+    echo 'Repo exists'
+  else
+    sh <(wget -qO - https://downloads.nordcdn.com/apps/linux/install.sh)
+  fi
 
   sudo groupadd --force nordvpn || true
   sudo usermod -aG nordvpn $USER || true
 }
 
 install_spotify() {
-  curl -sS https://download.spotify.com/debian/pubkey_6224F9941A8AA6D1.gpg | sudo gpg --dearmor --yes -o /etc/apt/trusted.gpg.d/spotify.gpg
-  echo "deb http://repository.spotify.com stable non-free" | sudo tee /etc/apt/sources.list.d/spotify.list
-  sudo apt-get update -y
-  sudo apt-get install -y spotify-client
+  if apt policy spotify-client 2>/dev/null
+  then
+    echo 'Repo exists'
+  else
+    curl -sS https://download.spotify.com/debian/pubkey_6224F9941A8AA6D1.gpg | sudo gpg --dearmor --yes -o /etc/apt/trusted.gpg.d/spotify.gpg
+    echo "deb http://repository.spotify.com stable non-free" | sudo tee /etc/apt/sources.list.d/spotify.list
+  fi
+
+  sudo apt update -y
+  sudo apt install -y spotify-client
 }
 
 install_golang() {
